Move functions goToGeo and goToChunk to GlobeBrowsingModule.
